+++ client/assets/scripts/custom/core/index.js
@@ -1,0 +1,4 @@
+export { default as initLoaders } from './loaders';
+export { default as initContact } from './contact';
+export { default as initNav } from './nav';
+export { default as initMap } from './map';