+++ client/assets/scripts/custom/core/map/index.js
@@ -1,0 +1,3 @@
+import initMap from './map';
+
+export default initMap;