+++ client/components/core/index.js
@@ -1,0 +1,9 @@
+import ApplicationRouter from './ApplicationRouter';
+import Head from './Head';
+import * as Misc from './misc';
+
+export {
+    ApplicationRouter,
+    Head,
+    Misc
+};