+++ client/components/core/misc/index.js
@@ -1,0 +1,9 @@
+import Contact from './Contact';
+import SectionLoader from './SectionLoader';
+import SocialIcons from './SocialIcons';
+
+export {
+    Contact,
+    SectionLoader,
+    SocialIcons
+};