+++ client/components/index.js
@@ -1,0 +1,5 @@
+import * as Core from './core';
+
+export {
+    Core
+};