+++ client/index.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import Config from './config';
+import { ApplicationRouter, Head } from './components/core';
 
 const {
     reactPresets: {
@@ -8,12 +9,19 @@ const {
     }
 } = Config;
 
+/**
+ * Core Application component
+ *
+ * @component
+ * @description Mounts the core application router.
+ */
 class Application extends Component {
     render() {
         return(
-            <h1>Hello, world!</h1>
+            <ApplicationRouter />
         )
     }
 }
 
+ReactDOM.render(<Head />, document.getElementById(renderDestinations.head));
 ReactDOM.render(<Application />, document.getElementById(renderDestinations.application));