+++ docs/changelog.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [1.0.1] - 2019-03-21
+### Added
+- Client-side configuration
+   - Config module
+- Browser router over application
+
 ## [1.0.0] - 2019-03-19
 ### Added
 - Created initial scaffolding